feat(librarian): decompose monolithic librarian into pure domain modules
